Adding endpoints to live tiles category
Adding additional endpoints that are reached out to from software that resides in the Live Tiles feature.  This software is owned by the Map of Internet Services team (moicore@microsoft.com) and customers have requested to have documentation updated to provide assurances that these endpoints are valid Microsoft endpoints. 
Endpoints have been added and the protocols used have been updated.
Please reach out with questions before merging as additional changes may need to be made and I am not sure of exactly how the ordering of the endpoints is desired to be set up here. 
Thank you!